+++ .devcontainer/devcontainer.json
@@ -8,10 +8,9 @@
 		"vscode": {
 			"extensions": [
 				"ms-python.python",
-				"jpotterm.simple-vim"
 			]
 		}
-	}
+	},
 
 	// Features to add to the dev container. More info: https://containers.dev/features.
 	// "features": {},
@@ -20,7 +19,7 @@
 	// "forwardPorts": [],
 
 	// Use 'postCreateCommand' to run commands after the container is created.
-	// "postCreateCommand": "pip3 install --user -r requirements.txt",
+	"postCreateCommand": ["pipx install poetry", "poetry install"],
 
 	// Configure tool-specific properties.
 	// "customizations": {},
